Seção 15: Implementando a Tela de Carrinho de Compras

+++ GeekShopping.Web/Models/CartHeaderViewModel.cs
@@ -5,5 +5,5 @@ public class CartHeaderViewModel
     public long Id { get; set; }
     public string? UserId { get; set; } = "";
     public string? CouponCode { get; set; } = "";
-    public double PurchaseAmout { get; set; }
+    public decimal PurchaseAmout { get; set; }
 }
